@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.3.0 <4.0.0'
+  sdk: '>=3.4.3 <4.0.0'
 
 
 dependencies:
@@ -20,32 +20,32 @@ dependencies:
   internet_connection_checker: ^1.0.0+1
   dartz: ^0.10.1
   get_it: ^7.6.7
-  dio: ^5.4.1
+  dio: ^5.4.3+1
   font_awesome_flutter: ^10.7.0
   flutter_svg: ^2.0.10+1
   flutter_staggered_grid_view: ^0.7.0
   youtube_player_flutter: ^9.0.0
   flutter_rating_bar: ^4.0.1
   readmore: ^2.2.0
-  intl: ^0.18.1
+  intl: ^0.19.0
   cached_network_image: ^3.3.1
   extended_wrap: ^0.1.5
   flutter_spinkit: ^5.2.1
-  persistent_bottom_nav_bar: ^5.0.2
+  persistent_bottom_nav_bar_v2: ^5.2.3
   lottie: ^3.1.0
-  geolocator: ^11.0.0
+  geolocator: ^12.0.0
   permission_handler: ^11.3.0
   syncfusion_flutter_charts: ^25.1.41+2
   flutter_html: 3.0.0-beta.2
-  url_launcher: ^6.2.6
+  url_launcher: ^6.3.0
   html: ^0.15.4
-  google_maps_flutter: ^2.6.1
+  google_maps_flutter: ^2.7.0
   like_button: ^2.0.5
   flutter_dotenv: ^5.1.0
-  device_info_plus: ^9.0.3
+  device_info_plus: ^9.1.2
   badges: ^3.1.2
   image_picker: ^1.1.2
-  fluttertoast: ^8.2.2
+  fluttertoast: ^8.2.6
   pin_code_fields: ^8.0.1
   animated_floating_buttons: ^0.0.2
 
